backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for photographer's personal site"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "bcrypt==4.3.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.124.4",
    "orjson>=3.11.5",
    "passlib[bcrypt]==1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
]
